replace all Es facade calls with Es::getClient() calls
