+++ app/Models/Wiki/Page.php
@@ -25,11 +25,11 @@ use App\Exceptions\GitHubNotFoundException;
 use App\Jobs\EsDeleteDocument;
 use App\Jobs\EsIndexDocument;
 use App\Libraries\Elasticsearch\BoolQuery;
+use App\Libraries\Elasticsearch\Es;
 use App\Libraries\OsuMarkdownProcessor;
 use App\Libraries\OsuWiki;
 use App\Libraries\Search\BasicSearch;
 use Carbon\Carbon;
-use Es;
 
 class Page
 {
@@ -155,12 +155,12 @@ class Page
         $params['body']['indexed_at'] = json_time(Carbon::now());
         $params['body']['version'] = static::VERSION;
 
-        return Es::index($params);
+        return Es::getClient()->index($params);
     }
 
     public function esDeleteDocument()
     {
-        return Es::delete(static::searchIndexConfig([
+        return Es::getClient()->delete(static::searchIndexConfig([
             'id' => $this->pagePath(),
             'client' => ['ignore' => 404],
         ]));
